IActionResult integrated

+++ HotelFinder.API/Controllers/HotelsController.cs
@@ -11,6 +11,7 @@ using System.Threading.Tasks;
 namespace HotelFinder.API.Controllers
 {
     [Route("api/[controller]")]
+    //api cont validationı kendi oto yapar 
     [ApiController]
     public class HotelsController : ControllerBase
     {
@@ -26,9 +27,11 @@ namespace HotelFinder.API.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        public List<Hotel> Get()
+        // public List<Hotel> Get()
+        public IActionResult Get()
         {
-            return _hotelService.GetAllHotels();
+            var hotels= _hotelService.GetAllHotels();
+            return Ok(hotels); //response 200 code ve body kısmında otelleri dön
         }
         /// <summary>
         /// Get Hotel By Id
@@ -37,9 +40,14 @@ namespace HotelFinder.API.Controllers
         /// <returns></returns>
 
         [HttpGet("{id}")]
-        public Hotel Get(int id)
+        public IActionResult Get(int id)
         {
-            return _hotelService.GetHotelById(id);
+            var hotel= _hotelService.GetHotelById(id);
+            if(hotel!=null)
+            {
+                return Ok(hotel); //200+data
+            }
+            return NotFound();//404
         }
 
         //create new hotel
@@ -49,9 +57,19 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPost]
-        public Hotel Post([FromBody]Hotel hotel)
+        public IActionResult Post([FromBody]Hotel hotel)
         {
-            return _hotelService.CreateHotel(hotel);
+            //Api controller kullanmasaydık böyle validation kontrol
+            //if(ModelState.IsValid)
+            //{
+            //    var createdHotel = _hotelService.CreateHotel(hotel);
+            //    return CreatedAtAction("Get", new { id = createdHotel.Id }, createdHotel);//201
+            //}
+            //return BadRequest(ModelState);//404
+
+            var createdHotel = _hotelService.CreateHotel(hotel);
+            //used get by id 
+            return CreatedAtAction("Get", new { id = createdHotel.Id }, createdHotel);//201 + data 
         }
 
         //Güncelle
@@ -61,9 +79,13 @@ namespace HotelFinder.API.Controllers
         /// <param name="hotel"></param>
         /// <returns></returns>
         [HttpPut]
-        public Hotel Put([FromBody]Hotel hotel)
+        public IActionResult Put([FromBody]Hotel hotel)
         {
-            return _hotelService.UpdateHotel(hotel);
+            if(_hotelService.GetHotelById(hotel.Id)!=null)
+            {
+                return Ok(_hotelService.UpdateHotel(hotel));//200 + data 
+            }
+            return NotFound();
         }
 
         /// <summary>
@@ -72,9 +94,14 @@ namespace HotelFinder.API.Controllers
         /// <param name="id"></param>
 
         [HttpDelete("{id}")]
-        public void Delete(int id )
+        public IActionResult Delete(int id )
         {
-            _hotelService.DeleteHotel(id);
+            if(_hotelService.GetHotelById(id)!=null)
+            {
+                _hotelService.DeleteHotel(id);
+                return Ok();//200
+            }
+            return NotFound();//404
         }
     }
 }
